Update project version and add implicit conversions
Updated the project version to 0.7.2 in Result.Api.csproj file. Also, implicit conversions are added to the ApiResult classes such that you can now implicitly convert an ApiResult to a string or its success model type, and likewise from a string or success model type to an ApiResult. This facilitates simpler and more intuitive usage of the ApiResult classes.