Update error handling in Result class and tests
Updated Result.Fail method to return an error message instead of an empty string. This change lets the consumer of the Result class handle failures more appropriately and makes debug easier. Additionally, a new unit test was added to validate this behavior. Also, the version in Result.csproj was bumped to 0.7.2 in preparation for the new release.